+++ pic16f17146-cold-plate-mplab-mcc.X/tempMonitor.c
@@ -162,7 +162,7 @@ void tempMonitor_loadResults(void)
             
             buffer += TS_OFFSET;
             buffer /= 10;
-            intTemp = buffer & 0xFF;
+            intTemp = (int8_t)(buffer & 0xFF);
             
             break;
         }
